Adds more specs or PBCore elements.
Adds specs for PBCore::Creator, PBCore::Creator::Creator, PBCore::Creator::Role,
PBCore::Coverage, PBCore::Coverage::Coverage, PBCore::Coverage::Type.

Example:

    xml = "<pbcoreCreator>
             <creator>Acme</creator>
			 <creatorRole source="foo.html">Producer</creatorRole>
		   </pbcoreCreator>"
    creator = PBCore::Creator.new
    creator.parse(xml)
    creator.creator # => returns an instance of PBCore::Creator::Creator
    creator.creator.value # => "Acme"
    creator.role # returns an instane of PBCore::Creator::Role
	creator.role.value # => "Producer"
	creator.role.source $ => "foo.html"

This PR also opts to use custom matchers instead of the shared spec. The shared
spec was becoming complex in getting it to work in all situations, and using it
was starting to required a bit of unintuitive setup. This will sacrific a bit of
DRYness for better readability, but i think it's a good change.

Also...
* Guards against calling #to_xml on pbcore element models that haven't defined a
  building block with class method .build_xml.
* Switches bin/console to use Pry instead of IRB.