Adds factories for all elements
This is a big update with 4 main goals:

1) Provide factories that produce 'complete' instances of each PBCore element,
   where 'complete' in this case meanse:
   a) all XML attributes have a value
   b) all child elements are present
   c) all child elements are also complete

   If a child element is repeatable, then 1 or more complete child elements
   are present (typically a random number beteen 1 and 3).

2) Refactor specs to use factories, and DRY things up a bit by using shared
   RSpec examples.

3) Use consistent naming conventions and folder structure for element classes,
   factories, and specs.

4) Remove WGBH-specific factories from the gem.

Most of the PBCore Element interfaces did not change except to make everything
more consistent. One exception though is the PBCore::Extension class where the
accessors `#value` and `#element` were changed to be `#extension_value` and
`#extension_element`. This is because `#value` is already used for the text
value within a PBCore::Element and was causing conflict, and because `#element`
is already a heavily used class method provided by the SAXMachine library, and
it is confusing to use it _also_ as an instance method that does something
very different.

Version bumped to 'pbcore-0.2.0' on rubygems.org.

Co-authored-by: foglabs <foglaboratories@gmail.com>